[Endpoint] increase result array size for RMW
In case a lot of operations in the queue are RMW every poll/signal
creates new IO requests. If we are out of requests we poll again.
Eventually we will run out of result arrays to poll to since we
only have enough for each request but don't take RMW operations
into account.

Signed-off-by: Jonas Pfefferle <pepperjo@japf.ch>